Moved most of interrupt handling from assembler handler to Interrupt_Wrapper
This commit simplifies the whole interrupt handling system by moving
code from the assembler handler to the Interrupt_Wrapper. This also
allows handling all pending interrupts of a priority in one call.

The cost of post-interrupt context-switch is now charged the
interrupted task. This may be better, as the interrupted task anyways
would be charged the cost of being context-switched back in again.