Searching time event queue from start to end
This reduces WCET for Set_Handler as the previous WCET included
searching the whole list and setting the hardware alarm timer. The
WCET should now be either searching the whole list or setting the
alarm timer, depending on which takes the longest time (which again
depends on the number of timers).

Time_Last is defined as Time'Last - 1 to ensure that all alarms set
are before Last_Alarm in the queue.