Fixed bug in reactivating idle clock after context switch
By removing Active_Clock from Thread_Descriptor a bug was introduced
meaning that the clock of the thread executing the idle loop would be
activated after a context switch even though it is not ready. This bug
is fixed by reintroducing Active_Clock.

Consider adding a dedicated idle thread as this would simplify the
system and provide deterministic release time when the system is
idling.